Split out common parts of the sample application
Also:
 * Initialise structs to zero
 * Get rid of global varible for Linux
 * Remove obsolete Windows cmake file
 * Implement station name command line argument for rt-kernel sample app
