--- cmake/Windows.cmake
@@ -1,22 +1,0 @@
-#********************************************************************
-#        _       _         _
-#  _ __ | |_  _ | |  __ _ | |__   ___
-# | '__|| __|(_)| | / _` || '_ \ / __|
-# | |   | |_  _ | || (_| || |_) |\__ \
-# |_|    \__|(_)|_| \__,_||_.__/ |___/
-#
-# www.rt-labs.com
-# Copyright 2018 rt-labs AB, Sweden.
-#
-# This software is dual-licensed under GPLv3 and a commercial
-# license. See the file LICENSE.md distributed with this software for
-# full license information.
-#*******************************************************************/
-
-set(GOOGLE_TEST_INDIVIDUAL TRUE)
-
-target_compile_options(profinet
-  PRIVATE
-  /WX
-  /D _CRT_SECURE_NO_WARNINGS
-  )
